requirements.txt
playwright
telethon
python-dotenv
